lab1

+++ backend/crud/ad.py
@@ -1,8 +1,7 @@
 from sqlalchemy.orm import Session
-from backend.models import Ad
+from backend.models import Ad, User, Role
 from backend.schemas.ad import AdCreate, AdUpdate
 from backend.core.storage import storage
-import os
 
 def get_ads(db: Session):
     return db.query(Ad).all()
@@ -28,10 +27,12 @@ def create_ad(db: Session, data: AdCreate, user_id: int):
 def get_user_ads(db: Session, user_id: int):
     return db.query(Ad).filter(Ad.user_id == user_id).all()
 
-def update_ad(db: Session, ad_id: int, data: AdUpdate, user_id: int):
-    ad = db.query(Ad).filter(Ad.id == ad_id, Ad.user_id == user_id).first()
+def update_ad(db: Session, ad_id: int, data: AdUpdate, user: User):
+    ad = db.query(Ad).filter(Ad.id == ad_id).first()
     if not ad:
         return None
+    if user.role != Role.admin and ad.user_id != user.id:
+        return False
     
     update_data = data.model_dump(exclude_unset=True)
     for field, value in update_data.items():
@@ -41,13 +42,15 @@ def update_ad(db: Session, ad_id: int, data: AdUpdate, user_id: int):
     db.refresh(ad)
     return ad
 
-def delete_ad(db: Session, ad_id: int, user_id: int):
-    ad = db.query(Ad).filter(Ad.id == ad_id, Ad.user_id == user_id).first()
+def delete_ad(db: Session, ad_id: int, user: User):
+    ad = db.query(Ad).filter(Ad.id == ad_id).first()
     if not ad:
         return None
+    if user.role != Role.admin and ad.user_id != user.id:
+        return False
     
     if ad.image:
-        storage.delete_file(ad.image)
+        storage.delete_file(ad.image) 
     
     db.delete(ad)
     db.commit()
